Extract quantity update related logic from `ProductRowViewModel` to `ProductStepperViewModel`.
