+++ WooCommerce/Classes/ViewRelated/Products/ProductSelector/ProductRow.swift
@@ -13,7 +13,7 @@ struct SimplifiedProductRow: View {
         HStack(alignment: .center) {
             Text(Localization.orderCountLabel)
             Spacer()
-            ProductStepper(viewModel: viewModel)
+            ProductStepper(viewModel: viewModel.stepperViewModel)
                 .renderedIf(viewModel.canChangeQuantity)
         }
     }
@@ -99,7 +99,7 @@ struct ProductRow: View {
                 .accessibilityLabel(viewModel.productAccessibilityLabel)
                 .accessibilityHint(accessibilityHint)
 
-                ProductStepper(viewModel: viewModel)
+                ProductStepper(viewModel: viewModel.stepperViewModel)
                     .renderedIf(viewModel.canChangeQuantity)
             }
         }
@@ -133,68 +133,6 @@ extension ProductRow {
     }
 }
 
-/// Represents a custom stepper.
-/// Used to change the quantity of the product in a `ProductRow`.
-///
-private struct ProductStepper: View {
-
-    /// View model to drive the view.
-    ///
-    @ObservedObject var viewModel: ProductRowViewModel
-
-    // Tracks the scale of the view due to accessibility changes
-    @ScaledMetric private var scale: CGFloat = 1
-
-    var body: some View {
-        HStack {
-            Button {
-                viewModel.decrementQuantity()
-            } label: {
-                Image(uiImage: .minusSmallImage)
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(height: Layout.stepperButtonSize * scale)
-            }
-            .disabled(viewModel.shouldDisableQuantityDecrementer)
-
-            Spacer()
-
-            Text(viewModel.quantity.description)
-
-            Spacer()
-
-            Button {
-                viewModel.incrementQuantity()
-            } label: {
-                Image(uiImage: .plusSmallImage)
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(height: Layout.stepperButtonSize * scale)
-            }
-            .disabled(viewModel.shouldDisableQuantityIncrementer)
-        }
-        .padding(Layout.stepperPadding * scale)
-        .frame(width: Layout.stepperWidth * scale)
-        .overlay(
-            RoundedRectangle(cornerRadius: Layout.stepperBorderRadius)
-                .stroke(Color(UIColor.separator), lineWidth: Layout.stepperBorderWidth)
-        )
-        .accessibilityElement(children: .ignore)
-        .accessibilityLabel("\(viewModel.name): \(Localization.quantityLabel)")
-        .accessibilityValue(viewModel.quantity.description)
-        .accessibilityAdjustableAction { direction in
-            switch direction {
-            case .decrement:
-                viewModel.decrementQuantity()
-            case .increment:
-                viewModel.incrementQuantity()
-            @unknown default:
-                break
-            }
-        }
-    }
-}
-
 private enum Layout {
     static let productImageSize: CGFloat = 48.0
     static let cornerRadius: CGFloat = 4.0
@@ -204,10 +142,6 @@ private enum Layout {
     static let stepperPadding: CGFloat = 11.0
     static let stepperWidth: CGFloat = 112.0
     static let checkImageSize: CGFloat = 24.0
-}
-
-private enum Localization {
-    static let quantityLabel = NSLocalizedString("Quantity", comment: "Accessibility label for product quantity field")
 }
 
 struct ProductRow_Previews: PreviewProvider {
