+++ WooCommerce/Classes/ViewRelated/Products/ProductSelector/ProductRowViewModel.swift
@@ -8,6 +8,8 @@ import WooFoundation
 final class ProductRowViewModel: ObservableObject, Identifiable {
     private let currencyFormatter: CurrencyFormatter
 
+    let stepperViewModel: ProductStepperViewModel
+
     /// Whether the product quantity can be changed.
     /// Controls whether the stepper is rendered.
     ///
@@ -257,38 +259,9 @@ final class ProductRowViewModel: ObservableObject, Identifiable {
 
     /// Quantity of product in the order
     ///
-    @Published private(set) var quantity: Decimal
-
-    /// Minimum value of the product quantity
-    ///
-    private let minimumQuantity: Decimal
-
-    /// Optional maximum value of the product quantity
-    ///
-    private let maximumQuantity: Decimal?
-
-    /// Whether the quantity can be decremented.
-    ///
-    var shouldDisableQuantityDecrementer: Bool {
-        if removeProductIntent != nil { // Allow decrementing below minimum quantity to remove product
-            return quantity < minimumQuantity
-        } else {
-            return quantity <= minimumQuantity
-        }
+    var quantity: Decimal {
+        stepperViewModel.quantity
     }
-
-    /// Whether the quantity can be incremented.
-    ///
-    var shouldDisableQuantityIncrementer: Bool {
-        guard let maximumQuantity else {
-            return false
-        }
-        return quantity >= maximumQuantity
-    }
-
-    /// Closure to run when the quantity is changed.
-    ///
-    var quantityUpdatedCallback: (Decimal) -> Void
 
     /// Closure to run when the quantity is decremented below the minimum quantity.
     ///
@@ -347,9 +320,12 @@ final class ProductRowViewModel: ObservableObject, Identifiable {
         self.stockStatus = .init(rawValue: stockStatusKey)
         self.stockQuantity = stockQuantity
         self.manageStock = manageStock
-        self.quantity = quantity
-        self.minimumQuantity = minimumQuantity
-        self.maximumQuantity = maximumQuantity
+        self.stepperViewModel = ProductStepperViewModel(quantity: quantity,
+                                                        name: name,
+                                                        minimumQuantity: minimumQuantity,
+                                                        maximumQuantity: maximumQuantity,
+                                                        quantityUpdatedCallback: quantityUpdatedCallback,
+                                                        removeProductIntent: removeProductIntent)
         self.canChangeQuantity = canChangeQuantity
         self.imageURL = imageURL
         self.hasParentProduct = hasParentProduct
@@ -360,7 +336,6 @@ final class ProductRowViewModel: ObservableObject, Identifiable {
         self.analytics = analytics
         self.numberOfVariations = numberOfVariations
         self.variationDisplayMode = variationDisplayMode
-        self.quantityUpdatedCallback = quantityUpdatedCallback
         self.removeProductIntent = removeProductIntent
         self.configure = configure
     }
@@ -555,30 +530,6 @@ final class ProductRowViewModel: ObservableObject, Identifiable {
     ///
     private func createAttributesText(from attributes: [VariationAttributeViewModel]) -> String {
         return attributes.map { $0.nameOrValue }.joined(separator: ", ")
-    }
-
-    /// Increment the product quantity.
-    ///
-    func incrementQuantity() {
-        if let maximumQuantity, quantity >= maximumQuantity {
-            return
-        }
-
-        quantity += 1
-
-        quantityUpdatedCallback(quantity)
-    }
-
-    /// Decrement the product quantity.
-    ///
-    func decrementQuantity() {
-        guard quantity > minimumQuantity else {
-            removeProductIntent?()
-            return
-        }
-        quantity -= 1
-
-        quantityUpdatedCallback(quantity)
     }
 
     func trackAddDiscountTapped() {
